simplify setup module
