--- py/desiutil/setup/__init__.py
@@ -1,16 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-# -*- coding: utf-8 -*-
-"""
-==============
-desiutil.setup
-==============
-
-This package contains code that might be useful in DESI setup.py files.
-"""
-from __future__ import absolute_import, division, print_function, unicode_literals
-from .desi_module import DesiModule
-from .desi_test import DesiTest
-from .desi_version import DesiVersion
-from .find_version_directory import find_version_directory
-from .get_version import get_version
-from .update_version import update_version
